Fix `ons_deaths` table update
This suggestion fixes part of the translation I missed in #20 and now
matches the intended behaviour of the earlier dataset definition.
Only dates and underlying causes that fall within pt_start_date and
pt_end_date are included in the dataset, otherwise NULL is returned.

I'm also deleting a comment because it would not work anymore without
applying the same case().when().then() logic.